+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Automatically download Terraform when it's not present
+- Automatically download Terragrunt when it's not present
 - Enabled FUSE async read 
 
 ## 0.2.1 - 2019-06-13